+++ tests/arbitrary-values.test.js
@@ -537,3 +537,41 @@ it('can explicitly specify type for percentage and length', () => {
     `)
   })
 })
+
+it('can use CSS variables as arbitrary values without `var()`', () => {
+  let config = {
+    content: [
+      {
+        raw: html`<div
+          class="w-[--width-var] bg-[--color-var] bg-[--color-var,#000] bg-[length:--size-var] text-[length:--size-var,12px]"
+        ></div>`,
+      },
+    ],
+    corePlugins: { preflight: false },
+    plugins: [],
+  }
+
+  let input = css`
+    @tailwind utilities;
+  `
+
+  return run(input, config).then((result) => {
+    expect(result.css).toMatchFormattedCss(css`
+      .w-\[--width-var\] {
+        width: var(--width-var);
+      }
+      .bg-\[--color-var\] {
+        background-color: var(--color-var);
+      }
+      .bg-\[--color-var\2c \#000\] {
+        background-color: var(--color-var, #000);
+      }
+      .bg-\[length\:--size-var\] {
+        background-size: var(--size-var);
+      }
+      .text-\[length\:--size-var\2c 12px\] {
+        font-size: var(--size-var, 12px);
+      }
+    `)
+  })
+})